bochs: use recursive mutexes with pthread
Summary:
Bochs use three types of locks depending of the configuration, SDL_Mutex, CriticalSection and pthread mutexes. The first two types are recursive but not the last one.

In some particular cases, like:
 - Calling a logging function from the simulator thread
 - Being signaled by the SIGALARM to display the IPS count
 - Calling a logging function to display the IPS
We are being deadlock because the logging function is using a lock before displaying its output.

So, just use `PTHREAD_MUTEX_RECURSIVE` to initialize pthread mutexes.

Test Plan: Test with two locks in a row in the same thread